fix test timeout
