+++ ddl/ingest/integration_test.go
@@ -19,17 +19,13 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/pingcap/tidb/ddl"
+	"github.com/pingcap/failpoint"
 	"github.com/pingcap/tidb/ddl/ingest"
-	"github.com/pingcap/tidb/ddl/testutil"
-	"github.com/pingcap/tidb/domain"
 	"github.com/pingcap/tidb/errno"
 	"github.com/pingcap/tidb/kv"
-	"github.com/pingcap/tidb/parser/model"
 	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/testkit"
 	"github.com/pingcap/tidb/tests/realtikvtest"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -98,61 +94,39 @@ func TestAddIndexIngestGeneratedColumns(t *testing.T) {
 	assertLastNDDLUseIngest(4)
 }
 
-func TestAddIndexIngestCancel(t *testing.T) {
-	store, dom := realtikvtest.CreateMockStoreAndDomainAndSetup(t)
+func TestIngestCopSenderErr(t *testing.T) {
+	store := testkit.CreateMockStore(t)
 	tk := testkit.NewTestKit(t, store)
 	tk.MustExec("use test;")
 	defer injectMockBackendMgr(t, store)()
 
-	tk.MustExec("create table t (a int, b int);")
-	tk.MustExec("insert into t (a, b) values (1, 1), (2, 2), (3, 3);")
-	defHook := dom.DDL().GetHook()
-	customHook := newTestCallBack(t, dom)
-	cancelled := false
-	customHook.OnJobRunBeforeExported = func(job *model.Job) {
-		if cancelled {
-			return
-		}
-		if job.Type == model.ActionAddIndex && job.SchemaState == model.StateWriteReorganization {
-			idx := testutil.FindIdxInfo(dom, "test", "t", "idx")
-			if idx == nil {
-				return
-			}
-			if idx.BackfillState == model.BackfillStateRunning {
-				tk2 := testkit.NewTestKit(t, store)
-				rs, err := tk2.Exec(fmt.Sprintf("admin cancel ddl jobs %d", job.ID))
-				assert.NoError(t, err)
-				assert.NoError(t, rs.Close())
-				cancelled = true
-			}
-		}
+	tk.MustExec("set @@global.tidb_ddl_reorg_worker_cnt = 1;")
+	tk.MustExec("create table t (a int primary key, b int);")
+	for i := 0; i < 4; i++ {
+		tk.MustExec(fmt.Sprintf("insert into t values (%d, %d);", i*10000, i*10000))
 	}
-	dom.DDL().SetHook(customHook)
-	tk.MustGetErrCode("alter table t add index idx(b);", errno.ErrCancelledDDLJob)
-	require.True(t, cancelled)
-	dom.DDL().SetHook(defHook)
-	ok, err := ingest.LitBackCtxMgr.CheckAvailable()
-	require.NoError(t, err)
-	require.True(t, ok)
+	tk.MustQuery("split table t between (0) and (50000) regions 5;").Check(testkit.Rows("4 1"))
+
+	require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/ddl/MockCopSenderError", "1*return"))
+	tk.MustExec("alter table t add index idx(a);")
+	require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/ddl/MockCopSenderError"))
+	tk.MustExec("admin check table t;")
+	rows := tk.MustQuery("admin show ddl jobs 1;").Rows()
+	//nolint: forcetypeassert
+	jobTp := rows[0][3].(string)
+	require.True(t, strings.Contains(jobTp, "ingest"), jobTp)
 }
 
-type testCallback struct {
-	ddl.Callback
-	OnJobRunBeforeExported func(job *model.Job)
-}
-
-func newTestCallBack(t *testing.T, dom *domain.Domain) *testCallback {
-	defHookFactory, err := ddl.GetCustomizedHook("default_hook")
-	require.NoError(t, err)
-	return &testCallback{
-		Callback: defHookFactory(dom),
-	}
-}
-
-func (c *testCallback) OnJobRunBefore(job *model.Job) {
-	if c.OnJobRunBeforeExported != nil {
-		c.OnJobRunBeforeExported(job)
-	}
+func TestAddIndexIngestPanicOnCopRead(t *testing.T) {
+	store := realtikvtest.CreateMockStoreAndSetup(t)
+	tk := testkit.NewTestKit(t, store)
+	tk.MustExec("use test;")
+	defer injectMockBackendMgr(t, store)()
+	require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/ddl/MockCopSenderPanic", "return(true)"))
+	tk.MustExec("create table t (a int, b int, c int, d int, primary key (a) clustered);")
+	tk.MustExec("insert into t (a, b, c, d) values (1, 1, 1, 1), (2, 2, 2, 2), (3, 3, 3, 3);")
+	tk.MustGetErrCode("alter table t add index idx(b);", errno.ErrReorgPanic)
+	require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/ddl/MockCopSenderPanic"))
 }
 
 func TestIngestPartitionRowCount(t *testing.T) {
